driver: add more tests, decompose commons
