+++ common/common/utils/set-utlis.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <set>
+
+namespace utils {
+template <typename T>
+static inline bool contains(std::set<T> set, T value) {
+    return set.find(value) != set.end();
+}
+}
