+++ driver/src/runtime/icd/kernel/CLKernel.cpp
@@ -1,4 +1,4 @@
-#include <common/common.hpp>
+#include <common/utils/common.hpp>
 #include <cstddef>
 #include <string>
 #include <utility>
