+++ driver/src/runtime/program/KernelArgumentInfoParser.cpp
@@ -1,4 +1,4 @@
-#include <common/common.hpp>
+#include <common/utils/common.hpp>
 
 #include "KernelArgumentInfoParser.h"
