+++ driver/src/runtime/runtime-command-queue.cpp
@@ -1,4 +1,4 @@
-#include <common/common.hpp>
+#include <common/utils/common.hpp>
 #include <iostream>
 
 #include "icd/CLCommandQueue.h"
