+++ driver/src/runtime/runtime-platform.cpp
@@ -1,4 +1,4 @@
-#include <common/common.hpp>
+#include <common/utils/common.hpp>
 #include <cstring>
 #include <iostream>
 #include <optional>
