+++ driver/test/unit/runtime/runtime-kernel-test.cpp
@@ -1,0 +1,45 @@
+#include <common/test/doctest.h>
+
+#include <runtime/runtime-commons.h>
+#include <vector>
+
+#include "runtime/icd/icd.h"
+#include "unit-test-common/test-commons.h"
+
+TEST_SUITE("Kernel API") {
+    TEST_CASE("clCreateKernel") {
+        SUBCASE("") {}
+    }
+
+    TEST_CASE("clCreateKernelsInProgram") {
+        SUBCASE("") {}
+    }
+
+    TEST_CASE("clRetainKernel") {
+        SUBCASE("") {}
+    }
+
+    TEST_CASE("clReleaseKernel") {
+        SUBCASE("") {}
+    }
+
+    TEST_CASE("clSetKernelArg") {
+        SUBCASE("") {}
+    }
+
+    TEST_CASE("clEnqueueNDRangeKernel") {
+        SUBCASE("") {}
+    }
+
+    TEST_CASE("clEnqueueTask") {
+        SUBCASE("") {}
+    }
+
+    TEST_CASE("clGetKernelArgInfo") {
+        SUBCASE("") {}
+    }
+
+    TEST_CASE("clGetKernelInfo") {
+        SUBCASE("") {}
+    }
+}
