+++ driver/test/unit/runtime/runtime-program-test.cpp
@@ -1,0 +1,49 @@
+#include <common/test/doctest.h>
+
+#include <runtime/runtime-commons.h>
+#include <vector>
+
+#include "runtime/icd/icd.h"
+#include "unit-test-common/test-commons.h"
+
+TEST_SUITE("Program API") {
+    TEST_CASE("clCreateProgramWithBinary") {
+        SUBCASE("") {}
+    }
+
+    TEST_CASE("clRetainProgram") {
+        SUBCASE("") {}
+    }
+
+    TEST_CASE("clReleaseProgram") {
+        SUBCASE("") {}
+    }
+
+    TEST_CASE("clBuildProgram") {
+        SUBCASE("") {}
+    }
+
+    TEST_CASE("clGetProgramInfo") {
+        SUBCASE("") {}
+    }
+
+    TEST_CASE("clCreateProgramWithSource") {
+        SUBCASE("") {}
+    }
+
+    TEST_CASE("clCompileProgram") {
+        SUBCASE("") {}
+    }
+
+    TEST_CASE("clLinkProgram") {
+        SUBCASE("") {}
+    }
+
+    TEST_CASE("clUnloadPlatformCompiler") {
+        SUBCASE("") {}
+    }
+
+    TEST_CASE("clUnloadCompiler") {
+        SUBCASE("") {}
+    }
+}
